% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrolldb.R
\name{color_samples}
\alias{color_samples}
\title{Color samples}
\usage{
color_samples(
  mzrolldb_file_path,
  unlabeled_samples_pattern,
  labeled_samples_pattern,
  unlabeled_color = c(0.929334998130799, 0.448004990816116, 0.18019400537014),
  labeled_color = c(0.122087001800537, 0.306293994188309, 0.916074991226196),
  other_color = c(0.5, 0.5, 0.5)
)
}
\arguments{
\item{mzrolldb_file_path}{file path to mzrolldb file}

\item{unlabeled_samples_pattern}{string pattern to identify samples in
representative unlabeled sample set.}

\item{labeled_samples_pattern}{string pattern to identify samples in
representative labeled sample set.}

\item{unlabeled_color}{color for all unlabeled samples.}

\item{labeled_color}{color for all labeled samples.}

\item{other_color}{color to color all samples that match to neither labeled
nor unlabeled samples.}
}
\value{
nothing (\code{invisible(0)}), the input file is mutated.
}
\description{
Color and reorder samples based on name agreement to provided regexes.
Note that this function mutates the mzrolldb file in place.
If you only wish to mutate a table without modifying a file, see
\code{diff_iso_color_samples()}.
}
