% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_diff_scores}
\alias{compute_diff_scores}
\title{Subset Diff Scores}
\usage{
compute_diff_scores(
  mzrolldb_file_path,
  isotopic_incorporation_scores,
  isotope_quant_measurement_type,
  sample_order,
  score_diff_threshold = 1.30103
)
}
\arguments{
\item{mzrolldb_file_path}{full path to mzrollDB file, which should already exist
and contain peak groups with isotopic envelopes.}

\item{isotope_quant_measurement_type}{Specific quant type used for evaluation.
For example, \code{peakArea}, \code{peakAreaTop}, \code{smoothedPeakArea}, or \code{peakIntensity}.}

\item{sample_order}{vector of all sample names, in desired order. Will be used for various sample display metrics.}

\item{score_diff_threshold}{threshold for determining signficance. By default, \code{-log10(0.05) = 1.30103}.}

\item{t_early_control_samples}{vector of sample names corresponding to \code{(t_early, control)} covariates.}

\item{t_early_treatment_samples}{vector of sample names corresponding to \code{(t_early, treatment)} covariates.}

\item{t_late_control_samples}{vector of sample names corresponding to \code{(t_late, control)} covariates.}

\item{t_late_treatment_samples}{vector of samples names corresponding to \code{(t_late, treatment)} covariates.}
}
\value{
design_matrix DataFrame consisting of 3 columns: \code{Measurement},
\code{Treatment}, and \code{Time}. The implied reference conditions are
understood to be \code{treatment} and \code{t_early}.
The provided columns are non-reference factor levels.
}
\description{
This step should follow isotopic incorporation, and as it requires as an input the
\code{isotopic_incorporation_scores} computed by \code{compute_isotopic_incorporation()}.

For each isotope associatd with each feature,
control and treatment subsets are determined. Each subset is separately subjected to a Welch t-test.
Then, the absolute score difference is taken of the \code{-log10(p-value)} from each test.
Mathematically, this is equivalent to a ratio of p-values.
If insufficient measurements are available to make a test, the score for one covariate subset is \code{0}.

This test may be useful for cases where more complex, integrated statistical tests might not be
computable. However, this function should probably be used in conjunction with the standard approach,
which constructs a linear model containing all samples/information (see \code{compute_time_emergent_diff_linear_model()}).
}
