% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_time_emergent_diff_linear_model}
\alias{compute_time_emergent_diff_linear_model}
\title{Compute Time-Emergent Diff linear model}
\usage{
compute_time_emergent_diff_linear_model(
  incorporation_subset,
  t_early_control_samples,
  t_early_treatment_samples,
  t_late_control_samples,
  t_late_treatment_samples,
  sample_order,
  p_val_interaction_threshold = 0.1,
  p_val_t_early_threshold = 0.05,
  p_val_t_late_threshold = 0.05,
  is_filter_M0_normalize = TRUE
)
}
\arguments{
\item{incorporation_subset}{list of isotope matrices that should be assessed for differences.}
}
\description{
Compute Time-Emergent Diff linear model
}
