% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_iso_scoring_functions.R
\name{diff_iso_all_isotopes_WelchTTest}
\alias{diff_iso_all_isotopes_WelchTTest}
\title{Welch's T-test for comparing isotopes in an envelope}
\usage{
diff_iso_all_isotopes_WelchTTest(
  isotope_matrix,
  condition_1_samples,
  condition_2_samples
)
}
\arguments{
\item{isotope_matrix}{DataFrame containing a \code{sample} column, and columns
for each isotope, usualy staring with \code{C12 PARENT}. Each row corresponds to a single
isotopic envelope for from a given sample. Usually generated by \code{mzkitcpp::ISO_isotope_matrices()}
followed by \code{to_iso_matrices()}.}

\item{condition_1_samples}{sample names corresponding to samples in
representative condition 1 sample set (usually control)}

\item{condition_2_samples}{sample names corresponding to samples in
representative condition 2 sample set (usually non-control)}
}
\value{
tibble of scores (by isotope) as \code{-log10(p_value)}, where higher scores are more significant (lower p-values).
that isotopic incorporation is evident based on the data in the \code{condition_1_samples}
and \code{condition_1_samples} in this \code{isotope_matrix}.
}
\description{
Every isotope in the Isotope Matrix is compared for two conditions
Using the Welch T-Test, and a set of p-values are returned.

Scoring:
A \code{-log10(p-value)} is computed for each kind of isotope.
If a p-value cannot be computed, the score is 0 (equivalent to p-value == 1)
}
