% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_iso_scoring_functions.R
\name{diff_iso_all_isotopes_WelchTTest_subset}
\alias{diff_iso_all_isotopes_WelchTTest_subset}
\title{Welch's T-test for comparing isotopes in an envelope}
\usage{
diff_iso_all_isotopes_WelchTTest_subset(
  isotope_matrix,
  subset_filters,
  condition_1_regex,
  condition_2_regex
)
}
\arguments{
\item{isotope_matrix}{DataFrame containing a \code{sample} column, and columns
for each isotope, usualy staring with \code{C12 PARENT}. Each row corresponds to a single
isotopic envelope for from a given sample. Usually generated by \code{mzkitcpp::ISO_isotope_matrices()}
followed by \code{to_iso_matrices()}.}

\item{subset_filters}{series of regexes that every sample that is to be compared must contain.
This vector input function as an AND filter (all surviving samples match every regex).}

\item{condition_1_regex}{used to determine sample names corresponding to samples in
representative condition 1 sample set (usually control)}

\item{condition_2_regex}{used to determine sample names corresponding to samples in
representative condition 2 sample set (usually non-control)}
}
\value{
tibble of scores (by isotope) as \code{-log10(p_value)}, where higher scores are more significant (lower p-values).
that isotopic incorporation is evident based on the data in the \code{condition_1_samples}
and \code{condition_1_samples} in this \code{isotope_matrix}.
}
\description{
A subset of samples is first extracted, based on a vector of string regexes.
Then, a string regex is used to further organize the remaining sample measurements
into either condition_1 or condition_2.
The results are passed along into \code{diff_iso_all_isotopes_WelchTTest()}.

Scoring:
A \code{-log10(p-value)} is computed for each kind of isotope.
If a p-value cannot be computed, the score is 0 (equivalent to p-value == 1)
}
