% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diff_iso_color_samples}
\alias{diff_iso_color_samples}
\title{Color and re-order samples}
\usage{
diff_iso_color_samples(
  samples_tbl,
  unlabeled_samples_pattern,
  labeled_samples_pattern,
  unlabeled_color = c(0.929334998130799, 0.448004990816116, 0.18019400537014),
  labeled_color = c(0.122087001800537, 0.306293994188309, 0.916074991226196),
  other_color = c(0.5, 0.5, 0.5)
)
}
\arguments{
\item{samples_tbl}{mzrollDB samples table, e.g. \code{PDB_sample_list()}.}

\item{unlabeled_samples_pattern}{string pattern to identify samples in
representative unlabeled sample set.}

\item{labeled_samples_pattern}{string pattern to identify samples in
representative labeled sample set.}

\item{unlabeled_color}{color for all unlabeled samples.}

\item{labeled_color}{color for all labeled samples.}

\item{other_color}{color to color all samples that match to neither labeled
nor unlabeled samples.}
}
\value{
sample table, to be saved to mzrollDB files.
}
\description{
Color samples based on labeled or unlabeled sample group,
also try to re-order them based on replicate number.
}
