% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_iso_rescoring.R
\name{diff_iso_conditions_rescore_and_label}
\alias{diff_iso_conditions_rescore_and_label}
\title{Re-score and label for isotopic incorporation and covariates of interest}
\usage{
diff_iso_conditions_rescore_and_label(
  mzrolldb_file,
  mzML_dir,
  unlabeled_samples_pattern,
  labeled_samples_pattern,
  experimental_design,
  incorporation_rescoring_function =
    metisotopes::diff_iso_m_plus_zero_fraction_WelchTTest,
  condition_rescoring_function = metisotopes::diff_iso_all_isotopes_WelchTTest,
  is_correct_natural_abundance = FALSE,
  rescore_suffix = "-rescored",
  incorporation_score_threshold = 1.30103,
  condition_score_threshold = 1.30103,
  verbose = TRUE
)
}
\arguments{
\item{mzrolldb_file}{mzrollDB file path, where \code{groupRank} column of
\code{peakgroups} will be updated after peak groups diff iso re-scoring.
This file should have been generated with isotopes.}

\item{mzML_dir}{local directory containing mzML or mzXML files.}

\item{unlabeled_samples_pattern}{string pattern to identify samples in
representative unlabeled sample set.}

\item{labeled_samples_pattern}{string pattern to identify samples in
representative labeled sample set.}

\item{experimental_design}{tibble describing various ways to subset the dataset.
Each subset will be evaluated for significance (only among labeled samples)
via application of the \code{condition_rescoring_function}.}

\item{incorporation_rescoring_function}{function for re-scoring peak groups based on
differential isotopic abundance.}

\item{condition_rescoring_function}{function for identifying individual isotopic
peak groups that are significance relative to one condition versus another.
These will be compared only among the labeled samples for peak groups
that exhibit isotopic incorporation.}

\item{is_correct_natural_abundance}{When determining isotope matrices, optionally
correct for natural abundance.}

\item{rescore_suffix}{name to be appended to the \code{original_mzrolldb_file}
(before the \code{.mzrollDB} file extension) in the renamed output file.}

\item{incorporation_score_threshold}{minimum value for diff iso scoring to label peak group.
Default value is \code{1.30103}, which corresponds to \code{-log10(0.05)}.}

\item{condition_score_threshold}{minimum value for a given peak group to be
considered rescored.
Default value is \code{1.30103}, which corresponds to \code{-log10(0.05)}.}

\item{verbose}{if \code{TRUE}, print additional messages to the console.}
}
\value{
(1) A long table listing out all peak groups where incorporation
   is thought to have been observed.
   This is the same set of peak groups that is used as a filter
   for labeled-only sample significance testing.
(2) A long table listing out all peak groups (including isotopic peak groups)
   where a significant difference by one condition is observed.

. Note also that a side-effect is the generation of a new mzrollDB file,
. named according to the \code{rescore_suffix} parameter, which is generated
. in the same directory as the input \code{original_mzrolldb_file}.
}
\description{
First, score peak groups based on the likelihood of isotopic incorporation.
This may work, for example, by comparing the fractional abundance of the [M+0] isotope
among unlabeled samples and labeled samples.
Then, among those peak groups that exhibit sufficient labeling incorporation,
Compare the differential abundance observed between different sample subsets
for each individual isotopic peak, for only the labeled samples.
The manner by which labeled samples may be divided into groups is specified
via the input parameter \code{experimental_design}.
In the initial 'isotopic incorporation' stage, isotope matrices are calculated
by using fractional abundance.
In the subsequent labeled stage, isotope matrices are calculated using the full
abundance.
A single argument, \code{is_correct_natural_abundance}, may be used to determine
if isotopes should correct natural abundance or not.
}
