% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_iso_scoring_functions.R
\name{diff_iso_emergent_significance}
\alias{diff_iso_emergent_significance}
\title{Emergent Significance}
\usage{
diff_iso_emergent_significance(design_matrix)
}
\arguments{
\item{design_matrix}{DataFrame consisting of 5 columns, Measurement,
Timet_early, Timet_late, Treatmenttreatment, Treatmentcontrol.}
}
\description{
Each sample should contain in variable 'Time' factors 't_early' and a 't_late' factor levels.
Each sample should contain in 'Treatment' factor 'control' and 'treatment' factor levels.

[1] Initial Linear Model

For each isotope, carry out the linear model

Measurement ~ Time + Treatment + (Time x Treatment)

Assume that the t_early and t_late samples are not from the same subjects - so it would be
inappropraite to use a linear mixed effects model.
Focus on the interaction term (Time x Treatment), looking for a significant p-value.

[2] Follow-up Games-Howell tests focusing on 't_late' vs 't_early' post-hoc comparisons
Demonstrate that t_early is not significant, while t_late is.

Scoring:
A \code{-log10(p-value)} is computed for each kind of isotope.
If a p-value cannot be computed, the score is 0 (equivalent to p-value == 1)
}
