% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_iso_scoring_functions.R
\name{diff_iso_m_plus_zero_fraction_WelchTTest}
\alias{diff_iso_m_plus_zero_fraction_WelchTTest}
\title{Welch's T-test for identifying isotopic incorporation.}
\usage{
diff_iso_m_plus_zero_fraction_WelchTTest(
  isotope_matrix,
  unlabeled_samples,
  labeled_samples
)
}
\arguments{
\item{isotope_matrix}{DataFrame containing a \code{sample} column, and columns
for each isotope, usualy staring with \code{C12 PARENT}. Each row corresponds to a single
isotopic envelope for from a given sample. Usually generated by \code{mzkitcpp::ISO_isotope_matrices()}
followed by \code{to_iso_matrices()}.}

\item{unlabeled_samples}{sample names corresponding to samples in
representative unlabeled sample set.}

\item{labeled_samples}{sample names corresponding to samples in
representative labeled sample set.}
}
\value{
score a value from 0 or 1, where values closer to 1 have greater evidence
that isotopic incorporation is evident based on the data in the \code{unlabeled_samples}
and \code{labeled_samples} in this \code{isotope_matrix}.
}
\description{
Compute the full Isotope Matrix, with no natural abundance correction,
and no exclusion of isotopic signal using reproducibility thresholds.
Then, compare the [M+0] fraction via Welch's t test.

Scoring:
Anything where the proportion of [M+0] is higher in the labeled samples than unlabeled gets a score of 0.
Anything with p > 0.05 gets a score of 0.
If a p-value could not be computed, the score is 0.
Otherwise, the score is (1 - pvalue) (also called the "confidence level").
}
