% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_iso_rescoring.R
\name{diff_iso_rescore}
\alias{diff_iso_rescore}
\title{Re-score Differential Abundance Isotopic Peak Groups.}
\usage{
diff_iso_rescore(
  mzrolldb_file,
  mzML_dir,
  rescoring_function,
  unlabeled_samples,
  labeled_samples,
  diff_iso_params,
  verbose = TRUE
)
}
\arguments{
\item{mzrolldb_file}{MzrollDB file path, where \code{groupRank} column of
\code{peakgroups} will be updated after peak groups diff iso re-scoring.}

\item{mzML_dir}{local directory containing mzML files.}

\item{rescoring_function}{function for re-scoring peak groups based on
differential isotopic abundance.}

\item{unlabeled_samples}{sample names corresponding to samples in
representative unlabeled sample set.}

\item{labeled_samples}{sample names corresponding to smaples in
representative labeled sample set.}

\item{diff_iso_params}{named list containing parameters that are used in the
construction of individual peak group isotopic matrices. Primarily used
by \code{mzkitcpp::ISO_isotope_matrices()} or \code{get_precomputed_iso_df()}.}

\item{verbose}{if \code{TRUE}, print additional messages to the console.}
}
\value{
a tibble comparing the original diff iso scores to the re-scored
values. The output tibble has columns \code{groupId}, \code{compoundId},
\code{groupRank} (the original score), \code{label}, \code{rescoredRank}, and
\code{diff}, which is \code{groupRank - rescoredRank}. This way, the original
score and the updated score can be compared directly.
}
\description{
Compare differences in the isotopic envelopes between two sample sets - a
representative unlabeled set, and a representative labeled set.
Override the 'groupRank' column in an mzrollDB file with the new score value.
}
