% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_iso_rescoring.R
\name{diff_iso_rescore_and_label}
\alias{diff_iso_rescore_and_label}
\title{Re-score and label Differential Abundance Isotopic Peak Groups.}
\usage{
diff_iso_rescore_and_label(
  original_mzrolldb_file,
  mzML_dir,
  rescoring_function,
  unlabeled_samples_pattern,
  labeled_samples_pattern,
  diff_iso_params,
  rescore_suffix = "-rescored",
  ms2_score_threshold = 0.7,
  rank_thresh = 1.30103,
  verbose = TRUE
)
}
\arguments{
\item{original_mzrolldb_file}{MzrollDB file path, where \code{groupRank} column of
\code{peakgroups} will be updated after peak groups diff iso re-scoring.}

\item{mzML_dir}{local directory containing mzML or mzXML files.}

\item{rescoring_function}{function for re-scoring peak groups based on
differential isotopic abundance.}

\item{unlabeled_samples_pattern}{string pattern to identify samples in
representative unlabeled sample set.}

\item{labeled_samples_pattern}{string pattern to identify samples in
representative labeled sample set.}

\item{diff_iso_params}{named list containing parameters that are used in the
construction of individual peak group isotopic matrices. Primarily used
by \code{mzkitcpp::ISO_isotope_matrices()} or \code{get_precomputed_iso_df()}.}

\item{rescore_suffix}{name to be appended to the \code{original_mzrolldb_file}
(before the \code{.mzrollDB} file extension) in the renamed output file.}

\item{ms2_score_threshold}{minimum value for MS2 score to label peak group.
This should be adjusted based on the scoring type, but for a cosine score
or similar default value of \code{0.7} is fine.}

\item{rank_thresh}{minimum value for diff iso scoring to label peak group.
Default value is \code{1.30103}, which corresponds to \code{-log10(0.05)}.}

\item{verbose}{if \code{TRUE}, print additional messages to the console.}
}
\value{
\code{invisible(0)}, as side effect, creates a new \code{mzrollDB} file
with the rescored values, named the same as the \code{original_mzrolldb_file}
except \code{rescore_suffix} is added to the name.
}
\description{
convenience function to copy an mzrollDB, re-score the values (via \code{diff_iso_rescore()}),
identify compounds of interest, label them, and save the rescored and labeled
peak groups into a new mzrollDB named from the initial mzrollDB with the
\code{-rescored.mzrollDB} suffix.
}
