% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_precomputed_iso_df}
\alias{get_precomputed_iso_df}
\title{Get mzrollDB isotopes as long table}
\usage{
get_precomputed_iso_df(
  iso_mzrolldb_file,
  isotope_quant_measurement_type = "smoothedPeakArea",
  is_fractional_abundance = FALSE,
  sample_order = NULL
)
}
\arguments{
\item{iso_mzrolldb_file}{full path of mzrollDB file.}

\item{isotope_quant_measurement_type}{column in \code{peak} table to use for quant.}

\item{is_fractional_abundance}{[FALSE] if \code{TRUE}, values are returned as fractions of
total isotopic envelope intensity for sample.}

\item{sample_order}{[NULL] if not \code{NULL}, and a list of sample names, will be used
to re-order the table, which may be useful for visualization downstream.}
}
\value{
iso_matrices_df
  Long table containing \code{groupId}, \code{groupMz}, \code{sample}, \code{isotope}, \code{measurement}.
  Note that \code{groupMz} will be a whole number if \code{is_collapse_nominal} is \code{TRUE}.
}
\description{
Utility function to pull out pre-computed isotopes (in an mzrollDB) as long table.
This works because peakdetector can support isotopic extraction, as part of
the mzrollDB generation step. Values are returned exactly as they were computed, though, with no adjustments.
Some adjustments are available as options to this function.
This function replaces / substitutes in for \code{mzkitcpp::ISO_isotope_matrices()}.

Note that this only operates on isotopic peaks - so if an mzrollDB contains a mixture
of isotopes and non-isotopes, this will only return back information about the
isotopic peaks
}
