% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_isotope_mzroll}
\alias{import_isotope_mzroll}
\title{Generate claman mzroll list from isotopes mzroll db file}
\usage{
import_isotope_mzroll(mzroll_db_path, method_tag, quant_type = "peakAreaTop")
}
\arguments{
\item{mzroll_db_path}{path to mzroll DB file,
or a pre-built romic triple_omic object with type tag 'mzroll'.
Callers of this function are responsible for passing in a valid input object
in that use case.  However, the object will still be tested via
\code{test_mzroll_list()}.}

\item{method_tag}{what method was run (for the purpose of matching
meta-data and aggregating).}
}
\value{
a **triple_omic** from **romic** containing three tibbles:
\itemize{
  \item{features: one row per unique analyte (defined by a unique
    groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{measurements: one row per peak (samples x peakgroups)}
  }
 And, a design list which tracks the variables in each table.
}
\description{
Generate process-ready mzroll from a single mzroll db file, with
corresponding method_tag (method name).
Note that this function assumes \code{only_identified} is \code{TRUE}.
The following rules are applied for peak groups:

Peak group filtering step:
(1) Remove any unidentified peak groups.
(2) Remove peak groups associated with isotopes - they are now accounted for via the envelope summation step.
(3) Remove any peak group from a non-Bookmarks search table if a Bookmark with the same name exists.
(4) Remove any peak group from a non-Bookmarks search table if multiple copies are found within a non-Bookmarks table.
(5) Remove any peak group from the Bookmarks search table if multiple copies are found within the Bookmarks table.

Peaks are re-quantified to correspond to the sum of the isotopic envelope, across samples.
}
