% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrolldb.R
\name{label_isotopes_by_top_hits}
\alias{label_isotopes_by_top_hits}
\title{Label mzrollDB file}
\usage{
label_isotopes_by_top_hits(mzrolldb_file_path, top_hits, sig_scores)
}
\arguments{
\item{mzrolldb_file_path}{file path to mzrolldb file}

\item{top_hits}{tibble consisting of at least columns \code{groupId}, \code{isotope}, and \code{groupRank} columns,
where \code{groupRank} refers to an isotope-specific score}

\item{sig_scores}{tibble consisting of at least columns \code{groupId}, \code{groupRank}, where \code{groupRank}
refers only to a score associated with evidence of isotopic incorporation.}
}
\description{
Based on a \code{top_hits} results tibble, update an mzrollDB file
by labeling peak groups appropriately.

Labeling follows the following rules:
label 'c' only - incorporation (only for parents), no sig isotopes
label 'a' incorporation + score diff > 1 - used for individual isotopes
label 'ca' sig isotopes and parents
}
