% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrolldb.R
\name{PDB_labeled_groups_summary}
\alias{PDB_labeled_groups_summary}
\title{Labeled groups summary}
\usage{
PDB_labeled_groups_summary(mzrolldb_file_path, label_regex = "c")
}
\arguments{
\item{label_regex}{regex for labels, will return all peak groups that match the regex}

\item{mzrolldb_file_path:}{file path to mzrolldb file}
}
\value{
formatted table based on high-scoring peakgroups with columns
\code{groupId}, \code{compoundName}, \code{adductName}, \code{groupMz}, \code{groupRt},
\code{ms2Score}, \code{groupRank}.
}
\description{
returns a formatted table, containing all peak groups that match the label_regex,
in descending order by groupRank. The groupRank column is frequently used to contain
rescored values, e.g. Welch T-test or other statistical tests.
}
