% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrolldb.R
\name{PDB_peakgroups}
\alias{PDB_peakgroups}
\title{Table of peakgroups from file}
\usage{
PDB_peakgroups(mzrolldb_file_path)
}
\arguments{
\item{mzrolldb_file_path:}{file path to mzrolldb file}
}
\value{
peakgroups table from mzrolldb file
}
\description{
Table of peakgroups from file
}
