% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrolldb.R
\name{PDB_peaks}
\alias{PDB_peaks}
\title{Table of peaks from file}
\usage{
PDB_peaks(mzrolldb_file_path)
}
\arguments{
\item{mzrolldb_file_path:}{file path to mzrolldb file}
}
\value{
peakgroups table from mzrolldb file
}
\description{
Table of peaks from file
}
