% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrolldb.R
\name{PDB_sample_list}
\alias{PDB_sample_list}
\title{Table of samples from file}
\usage{
PDB_sample_list(mzrolldb_file_path)
}
\arguments{
\item{mzrolldb_file_path:}{file path to mzrolldb file}
}
\value{
samples table from mzrolldb file
}
\description{
Table of samples from file
}
