% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakdetector.R
\name{peakdetector_add_CL_argument}
\alias{peakdetector_add_CL_argument}
\title{Add single command line parameter to args string}
\usage{
peakdetector_add_CL_argument(args, key, val)
}
\arguments{
\item{args}{Existing formatting arguments ready for peakdetector executable.}

\item{key}{parameter key, e.g. name from \code{params} named list.}

\item{val}{parameter value, e.g. stored value from \code{params} named list.}
}
\value{
updated \code{args} string, with data from single \code{(key, val)}
parameter formatted appropriately and appended to input \code{args}.
}
\description{
Add a single \code{(key, value)} parameter pair to \code{args} string.
Note that different kinds of keys require different handling in terms of their
formatting in the \code{args} string.
}
