% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakdetector.R
\name{peakdetector_add_params}
\alias{peakdetector_add_params}
\title{Add parameters to peakdetector command line}
\usage{
peakdetector_add_params(args, params)
}
\arguments{
\item{args}{Existing formatting arguments ready for peakdetector executable.}

\item{params}{Named list, containing command line options for peakdetector.}
}
\value{
updated \code{args} string, with data from \code{params} argument
formatted as a string and appended to input \code{args}.
}
\description{
Given a list of parameters, where names are command line option keys,
Generate a formatted string consumable by peakdetector executable.
Note that double-dashed parameters must follow single-dashed parameters,
so the formatted string is constructed to meet this requirement.
The keys are not checked, so if the peakdetector executable does not recognize a
key, it may cause an error or be ignored.
}
