% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakdetector.R
\name{peakdetector_add_rt_file}
\alias{peakdetector_add_rt_file}
\title{Add RT alignment info to peakdetector command line}
\usage{
peakdetector_add_rt_file(args, sample_directory)
}
\arguments{
\item{args}{Existing formatting arguments ready for peakdetector executable.}

\item{sample_directory}{Directory containing RT alignment files}
}
\value{
updated \code{args} string, with formatted samples string appended to
input \code{args}.
}
\description{
Given a folder containing one or more \code{.apts} or \code{.rt} files,
generate a formatted string consumable by peakdetector executable.
}
