% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakdetector.R
\name{peakdetector_add_samples}
\alias{peakdetector_add_samples}
\title{Add samples to peakdetector command line}
\usage{
peakdetector_add_samples(args, sample_directory)
}
\arguments{
\item{args}{Existing formatting arguments ready for peakdetector executable.}

\item{sample_directory}{Directory containing sample files.}
}
\value{
updated \code{args} string, with formatted samples string appended to
input \code{args}.
}
\description{
Given a folder containing one or more \code{.mzML} or \code{.mzXML} files,
generate a formatted string consumable by peakdetector executable.
}
