% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakdetector.R
\name{peakdetector_command_line}
\alias{peakdetector_command_line}
\title{Create peakdetector command line}
\usage{
peakdetector_command_line(
  peakdetector_executable,
  peakdetector_methods_folder,
  sample_directory,
  output_directory,
  params,
  is_save_ms2_scans = FALSE,
  is_ms1_search = FALSE
)
}
\arguments{
\item{peakdetector_executable}{absolute path to peakdetector executable.}

\item{peakdetector_methods_folder}{folder containing \code{.model} files, for
use with peak quality classification. By default, this should be
 \code{<peakdetector_executable>/Contents/Resources/methods/default.model}}

\item{sample_directory}{Directory containing set of \code{.mzML} or \code{.mzXML} files.}

\item{output_directory}{Folder where \code{peakdetector.mzrollDB} output file will be written.}

\item{params}{list of parameters consumed by peakdetector executable. See \code{peakdetector_default_parameters()}.}

\item{is_save_ms2_scans}{[FALSE] if \code{TRUE}, peak group ms2 scans are saved. Otherwise, they are not.}

\item{is_ms1_search}{[FALSE] if \code{TRUE}, do not include requirement to perform ms2-based peak detection.
When this option is \code{TRUE}, this will use the same parameters as are used in a MAVEN non-MS2 search.}
}
\value{
valid command line ready for \code{system(cmd)}
}
\description{
Creates a command line with appropriate arguments and values that can be properly
executed by \code{peakdetector} executable.
Arguments are re-ordered, with all double-dashed parameters following all
single-dashed parameters.
}
