% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakdetector.R
\name{peakdetector_default_isotope_parameters}
\alias{peakdetector_default_isotope_parameters}
\title{Default Peakdetector Isotope Parameters}
\usage{
peakdetector_default_isotope_parameters()
}
\value{
named list \code{isotope_params_list} with parameters, which should be modified
with the correct isotopes (and other parameters), and then used by \code{mzkitcpp::mzk_get_isotope_parameters()}.
}
\description{
Return a named list (e.g, \code{isotope_params_list}),containing default value of parameters from
the MAVEN GUI for an isotopes search. Note that the returned list does not actually specify any
isotopes, so this will not work without modification.
}
