% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakdetector.R
\name{peakdetector_default_parameters}
\alias{peakdetector_default_parameters}
\title{Default Peakdetector Parameters}
\usage{
peakdetector_default_parameters(spectral_library_file)
}
\arguments{
\item{spectral_library_file}{Full absolute path of spectral library, to be searched by peakdetector.}
}
\value{
named list \code{params} with parameters, ready for consumption of downstream functions,
e.g. \code{peakdetector_command_line()}.
}
\description{
Return a named list (e.g, \code{params}), based on an input \code{spectral_library_file}.
This function also serves as documentation, containing the default value of parameters from
the MAVEN GUI.
}
