% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakdetector.R
\name{peakdetector_metabolite_search_params}
\alias{peakdetector_metabolite_search_params}
\title{Metabolomics Parameters}
\usage{
peakdetector_metabolite_search_params(
  ms2MinNumMatches = 0,
  rtIsRequireRtMatch = 0,
  rtMatchTolerance = 1
)
}
\value{
encoded mzkitchen parameters string
}
\description{
Convenience Function to generate metabolomics parameters
}
