% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_iso_pipelines.R
\name{pipeline_diff_iso_conditions_search}
\alias{pipeline_diff_iso_conditions_search}
\title{Pipeline Diff Iso Search}
\usage{
pipeline_diff_iso_conditions_search(
  peakdetector_executable,
  peakdetector_methods_folder,
  peakdetector_params,
  sample_directory,
  output_directory,
  unscored_file_name,
  rescore_suffix,
  is_correct_natural_abundance,
  unlabeled_samples_pattern,
  labeled_samples_pattern,
  experimental_design,
  incorporation_score_threshold = 1.30103,
  condition_score_threshold = 1.30103,
  verbose = TRUE
)
}
\arguments{
\item{peakdetector_executable}{absolute path to compiled peakdetector executable.}

\item{peakdetector_methods_folder}{absolute path to peakdetector methods folder.}

\item{peakdetector_params}{list of peakdetector parameters.}

\item{sample_directory}{absolute path to directory containing all \code{mzML} files.}

\item{output_directory}{output directory for mzrollDB files.}

\item{unscored_file_name}{name of mzrollDB output file.}

\item{rescore_suffix}{added to file name in output directory for re-scored version
of output mzrollDB file.}

\item{is_correct_natural_abundance}{When determining isotope matrices, optionally
correct for natural abundance.}

\item{unlabeled_samples_pattern}{string pattern to identify samples in
representative unlabeled sample set.}

\item{labeled_samples_pattern}{string pattern to identify samples in
representative labeled sample set.}

\item{experimental_design}{tibble describing various ways to subset the dataset.
Each subset will be evaluated for significance (only among labeled samples)
via application of the \code{condition_rescoring_function}.}

\item{incorporation_score_threshold}{minimum value for diff iso scoring to label peak group.
Default value is \code{1.30103}, which corresponds to \code{-log10(0.05)}.}

\item{condition_score_threshold}{minimum value for a given peak group to be
considered rescored.
Default value is \code{1.30103}, which corresponds to \code{-log10(0.05)}.}

\item{verbose}{if \code{TRUE}, print additional messages to the console.}
}
\description{
Full Conditional Diff Iso pipeline - provide some configuration parameters,
search parameters, and iso parameters.
This function will create two files in the designated \code{output_directory}
: (1) a basic mzrollDB results file, and (2) a rescored mzrollDB file.
If there is any file named \code{peakdetector.mzrollDB} in the \code{output_directory},
it will be deleted.
This pipeline will use the conditional re-score to investigate any features
that are predicted to have undergone isotopic incorporation, and apply this to the supplied
experimental design structure to find overrepresentations.
}
