% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_iso_pipelines.R
\name{pipeline_diff_iso_search}
\alias{pipeline_diff_iso_search}
\title{Pipeline Diff Iso Search}
\usage{
pipeline_diff_iso_search(
  peakdetector_executable,
  peakdetector_methods_folder,
  peakdetector_params,
  sample_directory,
  output_directory,
  unscored_file_name,
  rescore_suffix,
  unlabeled_samples_pattern,
  labeled_samples_pattern,
  diff_iso_params,
  rank_thresh = 1.30103,
  verbose = TRUE
)
}
\arguments{
\item{peakdetector_executable}{absolute path to compiled peakdetector executable.}

\item{peakdetector_methods_folder}{absolute path to peakdetector methods folder.}

\item{peakdetector_params}{list of peakdetector parameters.}

\item{sample_directory}{absolute path to directory containing all \code{mzML} files.}

\item{output_directory}{output directory for mzrollDB files.}

\item{unscored_file_name}{name of mzrollDB output file.}

\item{rescore_suffix}{added to file name in output directory for re-scored version
of output mzrollDB file.}

\item{unlabeled_samples_pattern}{string pattern to identify samples in
representative unlabeled sample set.}

\item{labeled_samples_pattern}{string pattern to identify samples in
representative labeled sample set.}

\item{diff_iso_params}{named list containing parameters that are used in the
construction of individual peak group isotopic matrices. Primarily used
by \code{mzkitcpp::ISO_isotope_matrices()} or \code{get_precomputed_iso_df()}.}

\item{rank_thresh}{minimum value for diff iso scoring to label peak group.
Default value is \code{1.30103}, which corresponds to \code{-log10(0.05)}.}

\item{verbose}{if \code{TRUE}, print additional messages to the console.}
}
\description{
Full Diff Iso pipeline - provide some configuration parameters,
search parameters, and iso parameters.
Note that this pipeline does not involve conditional variables - only focused
On determining which samples show evidence of isotopic incorporation
(between labeled and unlabeled samples)
This function will create two files in the designated \code{output_directory}
: (1) a basic mzrollDB results file, and (2) a rescored mzrollDB file.
If there is any file named \code{peakdetector.mzrollDB} in the \code{output_directory},
it will be deleted.

Note that in this pipeline, isotopes are always compared using fractional values,
*even if the input parameter does not state as such*.
In other words, the parameter value is silently always set
\code{diff_iso_params[['diffIsoScoringFractionOfSampleTotal']] = TRUE}
}
