% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{requantify_to_envelope_sum}
\alias{requantify_to_envelope_sum}
\title{Generate quant values as envelope sum}
\usage{
requantify_to_envelope_sum(iso_df)
}
\arguments{
\item{iso_df:}{table of isotopic values, computed by \code{get_precomputed_iso_df()}.}
}
\value{
tibble of \code{envelopeSum} values for each \code{groupId}, e.g.
\code{groupId}, \code{sample}, and \code{envelopeSum} columns.
}
\description{
Using isotopic measurements table as generated by \code{get_precomputed_iso_df},
Generate new quantity values for each groupId based on summing across all isotopes.

Note that in that table, \code{groupId} actually refers to the parent groupId.

For downstream analysis, after using requantified values for \code{groupId},
all peak groups with a \code{parentGroupId} mapping to one of these \code{groupIds}
should probably be filtered out.
}
