% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_emergent_isotope_design_matrix}
\alias{to_emergent_isotope_design_matrix}
\title{Isotope-specific design matrix}
\usage{
to_emergent_isotope_design_matrix(
  isotope_matrix,
  isotope_name,
  t_early_control_samples,
  t_early_treatment_samples,
  t_late_control_samples,
  t_late_treatment_samples
)
}
\value{
norm_isotope_matrix normalized isotope matrix
}
\description{
Generate a linear-model ready design matrix from a peak group isotope_matrix,
specific isotope, and categorizations related to samples data.

The specific quantified isotope is renamed to 'Measurement'.

The quant values are also \code{log2}-transformed.
Any \code{NA}-values are filtered out.
}
