% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_iso_matrices}
\alias{to_iso_matrices}
\title{Utility function to convert long DataFrame into list of IsotopicMatrix tables.}
\usage{
to_iso_matrices(iso_matrices_df)
}
\arguments{
\item{iso_matrices_df}{Long dataframe containing columns
.  \code{sample}, \code{isotope}, and \code{measurement}, which would typically
  be produced by \code{mzkitcpp::ISO_isotope_matrices()}.}
}
\value{
iso_matrices_reshaped: a named list, where entries are isotope matrix
data frames (columns: \code{sample}, followed by all individual isotope names;
rows: every sample's isotopic envelope).  Entries are named by \code{groupId},
and can be accessed, for example, with \code{groupId = 1234} via \code{isotope_matrices_reshaped$`1234`}
}
\description{
Designed to convert the output of \code{mzkitcpp::ISO_isotope_matrices()} to
a list of tibbles, each of which corresponds to a single isotope matrix.
this isotope matrix is an R version of the \code{maven_core IsotopeMatrix} class.
}
