% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_M0_normalized_isotope_matrix}
\alias{to_M0_normalized_isotope_matrix}
\title{Normalize isotope matrix to sample \code{[M+0]}}
\usage{
to_M0_normalized_isotope_matrix(isotope_matrix)
}
\value{
norm_isotope_matrix normalized isotope matrix
}
\description{
Divide all non-M0 isotopes by M0 isotope abundance.
\code{NA} values are retained as \code{NA}.
Peaks are re-quantified to correspond to the sum of the isotopic envelope, across samples.
}
