% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_isotopic_incorporation}
\alias{compute_isotopic_incorporation}
\title{Determine Isotopic Incorporation}
\usage{
compute_isotopic_incorporation(
  mzrolldb_file_path,
  isotope_quant_measurement_type,
  t_early_control_samples,
  t_early_treatment_samples,
  t_late_control_samples,
  t_late_treatment_samples,
  sample_order
)
}
\value{
design_matrix DataFrame consisting of 3 columns: \code{Measurement},
\code{Treatment}, and \code{Time}. The implied reference conditions are
understood to be \code{treatment} and \code{t_early}.
The provided columns are non-reference factor levels.
}
\description{
Generate a linear-model ready design matrix from a peak group isotope_matrix,
specific isotope, and categorizations related to samples data.

The specific quantified isotope is renamed to 'Measurement'.

The quant values are also \code{log2}-transformed.
Any \code{NA}-values are filtered out.
}
