% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_iso_scoring_functions.R
\name{diff_iso_emergent_significance}
\alias{diff_iso_emergent_significance}
\title{Time-Emergent Differential Abundance}
\usage{
diff_iso_emergent_significance(
  isotope_matrix,
  is_M0_normalize,
  t_early_control_samples,
  t_late_control_samples,
  t_early_treatment_samples,
  t_late_treatment_samples
)
}
\arguments{
\item{isotope_matrix}{DataFrame containing isotope measurements.}

\item{is_M0_normalize}{if \code{TRUE}, every isotope value is divided by the [M+0] value from its respective sample.}

\item{t_early_control_samples}{sample names corresponding to \code{(Time=t_early, Treatment=control)}}

\item{t_late_control_samples}{sample names corresponding to \code{(Time=t_late, Treatment=control)}}

\item{t_early_treatment_samples}{sample names corresponding to \code{(Time=t_early, Treatment=treatment)}}

\item{t_late_treatment_samples}{sample names corresponding to \code{(Time=t_late, Treatment=treatment)}}
}
\value{
data frame containing p-value summaries of results.
to convert to an overall score, the \code{p_val_interaction} should be BELOW a threshold parameter,
(p-value significant)
the \code{p_val_early_diff} should be ABOVE a threshold parameter (p-value not significant)
the \code{p_val_late_diff} should be BELOW a threshold parameter (p-value significant)

For example,
let isotope == [M+2], p_val_interaction = 0.0715, p_val_early_diff = 0.693, and p_val_late_diff = 0.0197.

let p_val_interaction_thresh = 0.10, p_val_early_thresh = 0.05, p_val_late_thresh = 0.05.

This example passes because
(1) p_val_interaction = 0.0715 < p_val_interaction_thresh = 0.10 --> PASSES INTERACTION TERM TEST
(2) p_val_early_thresh = 0.05 > p_val_early_diff = 0.693 --> PASSES T_EARLY TEST
(3) p_val_late_thresh = 0.05 < p_val_late_diff = 0.0197 --> PASSES T_LATE_TEST

OVERALL RESULT: this isotope shows time-emergent differential abundance.

Intuitively,
(1) Application of the treatment at a late time point revealed a significant difference compared
.    to the control at the initial time point, the control alone, or the effect of time alone.
(2) Application of this treatment was not significant at the initial time point (t_early).
(3) Application of this treatment was verified at the later time-point.
}
\description{
Each sample should contain a covariate 'Time', with implied reference factor level 't_early' and explicit factor level 't_late'.
Each sample should contain a covariate 'Treatment', with implied factor level 'control' and explicit factor level 'treatment'.

For each isotope, carry out the linear model

Measurement ~ Time + Treatment + (Time x Treatment)

Assume that the t_early and t_late samples are not from the same subjects - so it would be
inappropriate to use a linear mixed effects model.
Focus on the interaction term (Time x Treatment), looking for a significant p-value.

Significance must pass two p-value thresholds, one for the interaction term (must be less than this value),
another for the comparison of condition vs treatment for the late time points (must be greater than this value).

[2] Follow-up Games-Howell tests focusing on 't_late' vs 't_early' post-hoc comparisons.

Ultimately,the difference between treatment and control should
(1) SHOULD have a significant \code{p_val_interaction} value,
(2) SHOULD NOT have a significant \code{p_val_early_diff} value,
(3) SHOULD have a significant \code{p_val_late_diff} value.

Significance can be determined downstream by the consideration of these 3 p-values.
Scores will often take the form of \code{-log10(p-value)}, where the final score should
probably be the \code{p_val_late_diff} value.
}
